#314
- side-page: добавлены e2e-тесты
